+++ dataweaver/apps/web/src/components/scopes/atlas/components/in_front_of_canvas/index.tsx
@@ -1,0 +1,11 @@
+import { Controls } from './controls';
+import { Tools } from './tools';
+
+export const InFrontOfTheCanvas = () => {
+  return (
+    <>
+      <Controls />
+      <Tools />
+    </>
+  );
+};
