+++ dataweaver/apps/web/src/components/scopes/atlas/components/in_front_of_canvas/tools.module.scss
@@ -1,0 +1,17 @@
+.container {
+  position: absolute;
+  top: 200px;
+  right: 16px;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  align-items: center;
+  width: 58px;
+  padding: 8px;
+  pointer-events: auto;
+  background: rgb(var(--color-control-surface));
+  border-radius: 29px;
+  box-shadow:
+    0 1px 3px 1px rgb(var(--color-shadow) / 15%),
+    0 1px 2px rgb(var(--color-shadow) / 30%);
+}
