+++ dataweaver/apps/web/src/components/scopes/atlas/components/in_front_of_canvas/tools.tsx
@@ -1,0 +1,70 @@
+import { useEditor, useValue } from 'tldraw';
+import { Button } from '~/components/elements/button';
+import { IconCursor } from '~/components/primitives/icons/cursor';
+import { IconHand } from '~/components/primitives/icons/hand';
+import { IconInsertText } from '~/components/primitives/icons/insert_text';
+import s from './tools.module.scss';
+
+const TOOLS = {
+  select: { label: 'Select', Icon: IconCursor },
+  hand: { label: 'Pan', Icon: IconHand },
+  text: { label: 'Text', Icon: IconInsertText },
+} as const;
+
+type ToolName = keyof typeof TOOLS;
+
+/** Type guard to ensure a given string is a valid ToolName. */
+const isToolName = (name: string): name is ToolName => name in TOOLS;
+
+const BUTTON_TOOL_COLOR_SCHEME_SELECTED = {
+  base: 'var(--color-control-accent)',
+  'base-hover': 'var(--color-control-accent)',
+  content: 'var(--color-control-accent-content)',
+  'content-hover': 'var(--color-control-accent-content)',
+};
+
+const BUTTON_TOOL_COLOR_SCHEME_INACTIVE = {
+  base: 'transparent',
+  'base-hover': 'var(--color-control-surface-hover)',
+  content: 'var(--color-control-content)',
+  'content-hover': 'var(--color-control-content)',
+};
+
+/**
+ * Editor-bound wrapper rendered through tldraw's `InFrontOfTheCanvas` slot so
+ * it can read live tool / zoom state via `useEditor`.
+ */
+export const Tools = () => {
+  const editor = useEditor();
+
+  const tool = useValue('tool', () => editor.getCurrentToolId(), [editor]);
+
+  // Here we map the tldraw tool to one of our supported ones. If it's not found
+  // we fallback to default 'select' tool
+  const activeToolName: ToolName = isToolName(tool) ? tool : 'select';
+
+  return (
+    <div
+      className={s.container}
+      role="toolbar"
+      aria-orientation="vertical"
+      aria-label="Tools"
+    >
+      {Object.entries(TOOLS).map(([name, tool]) => (
+        <Button
+          key={name}
+          icon={tool.Icon}
+          size="large"
+          aria-label={tool.label}
+          aria-pressed={activeToolName === name}
+          colorScheme={
+            activeToolName === name
+              ? BUTTON_TOOL_COLOR_SCHEME_SELECTED
+              : BUTTON_TOOL_COLOR_SCHEME_INACTIVE
+          }
+          onClick={() => editor.setCurrentTool(name)}
+        />
+      ))}
+    </div>
+  );
+};
